changes based on github comments
